+++ src/RhinoInside.Revit.GH/Types/Annotations/Revision.cs
@@ -11,5 +11,37 @@ namespace RhinoInside.Revit.GH.Types
 
     public Revision() { }
     public Revision(ARDB.Revision element) : base(element) { }
+
+    public string RevisionNumber => Value?.get_Parameter(ARDB.BuiltInParameter.PROJECT_REVISION_REVISION_NUM)?.AsString();
+    public int? SequenceNumber => Value?.get_Parameter(ARDB.BuiltInParameter.PROJECT_REVISION_SEQUENCE_NUM)?.AsInteger();
+
+    public string Description
+    {
+      get => Value?.Description;
+      set { if (value is object && Value is ARDB.Revision revision && revision.Description != value) revision.Description = value; }
+    }
+
+    public string RevisionDate
+    {
+      get => Value?.RevisionDate;
+      set { if (value is object && Value is ARDB.Revision revision && revision.RevisionDate != value) revision.RevisionDate = value; }
+    }
+
+    public string IssuedBy
+    {
+      get => Value?.IssuedBy;
+      set { if (value is object && Value is ARDB.Revision revision && revision.IssuedBy != value) revision.IssuedBy = value; }
+    }
+
+    public string IssuedTo
+    {
+      get => Value?.IssuedTo;
+      set { if (value is object && Value is ARDB.Revision revision && revision.IssuedTo != value) revision.IssuedTo = value; }
+    }
+    public bool? Issued
+    {
+      get => Value?.Issued;
+      set { if (value is object && Value is ARDB.Revision revision && revision.Issued != value) revision.Issued = value.Value; }
+    }
   }
 }